# Conflicts:
#	newper_admin/src/main/java/com/newper/service/OrderService.java

+++ newper_admin/src/main/java/com/newper/repository/ShopProductOptionRepo.java
@@ -4,5 +4,5 @@ import com.newper.entity.ShopProductOption;
 import com.newper.entity.Spec;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShopProductOptionRepo extends JpaRepository<ShopProductOption, Integer> {
+public interface ShopProductOptionRepo extends JpaRepository<ShopProductOption, Long> {
 }
